Add files via upload
Copy the file ex1.asm to ex2.asm. Modify the program in ex2.asm to make it work on multiple inputs. In particular, it should repeatedly ask for width and height values, and print the corresponding area and perimeter, until the user enters the value 0 for width. At that point, the program should terminate. Here's an execution scenario:

Enter width (integer):
2
Enter height (integer):
4
Rectangle's area is 8
Its perimeter is 12
Enter width (integer):
5
Enter height (integer):
6
Rectangle's area is 30
Its perimeter is 22
Enter width (integer):
0
Essentially, this exercise involves introducing a loop around the main code of Exercise 1. As long as width is not equal to 0, the program repeats by looping back to the top of the loop. If width is 0, the program breaks out of the loop. The key instructions for forming such loops are: beq, bne and j. You may need one or more of these, depending on how you construct your loop.